refactor: move state interface to api
