+++ engine/src/main/java/io/camunda/zeebe/streamplatform/ProcessingStateMachine.java
@@ -9,11 +9,6 @@ package io.camunda.zeebe.streamplatform;
 
 import io.camunda.zeebe.db.TransactionContext;
 import io.camunda.zeebe.db.ZeebeDbTransaction;
-import io.camunda.zeebe.streamplatform.api.EmptyProcessingResult;
-import io.camunda.zeebe.streamplatform.api.ProcessingResult;
-import io.camunda.zeebe.streamplatform.api.ProcessingResultBuilder;
-import io.camunda.zeebe.streamplatform.api.RecordProcessor;
-import io.camunda.zeebe.streamplatform.api.TypedRecord;
 import io.camunda.zeebe.engine.metrics.StreamProcessorMetrics;
 import io.camunda.zeebe.engine.processing.streamprocessor.RecordValues;
 import io.camunda.zeebe.logstreams.impl.Loggers;
@@ -28,7 +23,12 @@ import io.camunda.zeebe.scheduler.future.ActorFuture;
 import io.camunda.zeebe.scheduler.retry.AbortableRetryStrategy;
 import io.camunda.zeebe.scheduler.retry.RecoverableRetryStrategy;
 import io.camunda.zeebe.scheduler.retry.RetryStrategy;
-import io.camunda.zeebe.streamplatform.state.MutableLastProcessedPositionState;
+import io.camunda.zeebe.streamplatform.api.EmptyProcessingResult;
+import io.camunda.zeebe.streamplatform.api.ProcessingResult;
+import io.camunda.zeebe.streamplatform.api.ProcessingResultBuilder;
+import io.camunda.zeebe.streamplatform.api.RecordProcessor;
+import io.camunda.zeebe.streamplatform.api.TypedRecord;
+import io.camunda.zeebe.streamplatform.api.state.MutableLastProcessedPositionState;
 import io.camunda.zeebe.util.buffer.BufferUtil;
 import io.camunda.zeebe.util.exception.RecoverableException;
 import io.camunda.zeebe.util.exception.UnrecoverableException;
