+++ engine/src/main/java/io/camunda/zeebe/streamplatform/api/state/MutableLastProcessedPositionState.java
@@ -5,9 +5,9 @@
  * Licensed under the Zeebe Community License 1.1. You may not use this file
  * except in compliance with the Zeebe Community License 1.1.
  */
-package io.camunda.zeebe.streamplatform.state;
+package io.camunda.zeebe.streamplatform.api.state;
 
-public interface LastProcessedPositionState {
+public interface MutableLastProcessedPositionState extends LastProcessedPositionState {
 
-  long getLastSuccessfulProcessedRecordPosition();
+  void markAsProcessed(final long position);
 }
