+++ engine/src/test/java/io/camunda/zeebe/engine/util/StreamProcessingComposite.java
@@ -24,7 +24,7 @@ import io.camunda.zeebe.scheduler.ActorScheduler;
 import io.camunda.zeebe.scheduler.future.ActorFuture;
 import io.camunda.zeebe.streamplatform.StreamProcessor;
 import io.camunda.zeebe.streamplatform.StreamProcessorListener;
-import io.camunda.zeebe.streamplatform.state.MutableLastProcessedPositionState;
+import io.camunda.zeebe.streamplatform.api.state.MutableLastProcessedPositionState;
 import java.util.Optional;
 import java.util.concurrent.Callable;
